rename package service to services
